+++ src/main/java/org/eclipselabs/garbagecat/domain/jdk/CmsSerialOldEvent.java
@@ -145,9 +145,9 @@ public class CmsSerialOldEvent extends CmsIncrementalModeCollector implements Bl
     private String logEntry;
 
     /**
-     * The elapsed clock time for the GC event in micriseconds (rounded).
+     * The elapsed clock time for the GC event in microseconds (rounded).
      */
-    private int duration;
+    private long duration;
 
     /**
      * The time when the GC event started in milliseconds after JVM startup.
@@ -403,7 +403,7 @@ public class CmsSerialOldEvent extends CmsIncrementalModeCollector implements Bl
         this.logEntry = logEntry;
     }
 
-    public int getDuration() {
+    public long getDuration() {
         return duration;
     }
