+++ src/main/java/org/eclipselabs/garbagecat/domain/jdk/G1RemarkEvent.java
@@ -85,7 +85,7 @@ public class G1RemarkEvent extends G1Collector implements BlockingEvent, Paralle
     /**
      * The elapsed clock time for the GC event in microseconds (rounded).
      */
-    private int duration;
+    private long duration;
 
     /**
      * The time when the GC event started in milliseconds after JVM startup.
@@ -154,7 +154,7 @@ public class G1RemarkEvent extends G1Collector implements BlockingEvent, Paralle
         return logEntry;
     }
 
-    public int getDuration() {
+    public long getDuration() {
         return duration;
     }
