+++ src/main/java/org/eclipselabs/garbagecat/domain/jdk/SerialNewEvent.java
@@ -87,7 +87,7 @@ public class SerialNewEvent extends SerialCollector
     /**
      * The elapsed clock time for the GC event in microseconds (rounded).
      */
-    private int duration;
+    private long duration;
 
     /**
      * The time when the GC event started in milliseconds after JVM startup.
@@ -194,7 +194,7 @@ public class SerialNewEvent extends SerialCollector
         return logEntry;
     }
 
-    public int getDuration() {
+    public long getDuration() {
         return duration;
     }
