+++ src/main/java/org/eclipselabs/garbagecat/domain/jdk/unified/UnifiedCmsInitialMarkEvent.java
@@ -70,7 +70,7 @@ public class UnifiedCmsInitialMarkEvent extends CmsCollector
     /**
      * The elapsed clock time for the GC event in microseconds (rounded).
      */
-    private int duration;
+    private long duration;
 
     /**
      * The time when the GC event started in milliseconds after JVM startup.
@@ -141,7 +141,7 @@ public class UnifiedCmsInitialMarkEvent extends CmsCollector
     }
 
     /**
-     * Alternate constructor. Create CMS Initial Mark from values.
+     * Alternate constructor. Create detail logging event from values.
      * 
      * @param logEntry
      *            The log entry for the event.
@@ -160,7 +160,7 @@ public class UnifiedCmsInitialMarkEvent extends CmsCollector
         return logEntry;
     }
 
-    public int getDuration() {
+    public long getDuration() {
         return duration;
     }
