+++ src/main/java/org/eclipselabs/garbagecat/domain/jdk/unified/UnifiedParallelScavengeEvent.java
@@ -86,7 +86,7 @@ public class UnifiedParallelScavengeEvent extends ParallelCollector implements U
     /**
      * The elapsed clock time for the GC event in microseconds (rounded).
      */
-    private int duration;
+    private long duration;
 
     /**
      * The time when the GC event started in milliseconds after JVM startup.
@@ -238,7 +238,7 @@ public class UnifiedParallelScavengeEvent extends ParallelCollector implements U
         return logEntry;
     }
 
-    public int getDuration() {
+    public long getDuration() {
         return duration;
     }
